Move dinosaur and bubble functionality to separate files
Export as modules and require in inedx
